[fetch] rename index to fetch-proxy.
